Restructure project
Added an App struct to represent db and router, and relevant methods.
