+++ personController_test.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"testing"
+)
+
+func TestGetPersons(t *testing.T) {
+	// req, err := http.NewRequest("GET", "/api/persons", nil)
+	// if err != nil {
+	// 	t.Fatal(err)
+	// }
+	// fmt.Printf("%v\n%v\n", req, err)
+	// rr := httptest.NewRecorder()
+	// fmt.Println(rr)
+	// handler := http.HandlerFunc(GetPersons)
+	// handler.ServeHTTP(rr, req)
+	// if status := rr.Code; status != http.StatusOK {
+	// 	t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusOK)
+	// }
+
+	// TODO Check response body
+	// expected := ``
+	// if rr.Body.String() != expected {
+	// 	t.Errorf("handler returned unexpected body: got %v want %v", rr.Body.String(), expected)
+	// }
+}
